+++ tests/unit/devices/WH65.spec.ts
@@ -21,6 +21,7 @@ const dataReport1 = {
   "rainratein": "0.03",
   "eventrainin": "0.4",
   "hourlyrainin": "0.5",
+  "last24hrainin": "0.450",
   "dailyrainin": "0.000",
   "weeklyrainin": "0.272",
   "monthlyrainin": "0.819",
@@ -100,6 +101,7 @@ configs.forEach(config => {
       expect(device.rainRate.service.characteristics[0].value).to.equal("Rain Rate");
       expect(device.eventRain.service.characteristics[0].value).to.equal("Rain Event Total");
       expect(device.hourlyRain.service.characteristics[0].value).to.equal("Rain Hourly Total");
+      expect(device.last24hRain.service.characteristics[0].value).to.equal("Rain Last 24h Total");
       expect(device.dailyRain.service.characteristics[0].value).to.equal("Rain Daily Total");
       expect(device.weeklyRain.service.characteristics[0].value).to.equal("Rain Weekly Total");
       expect(device.monthlyRain.service.characteristics[0].value).to.equal("Rain Monthly Total");
@@ -125,6 +127,7 @@ configs.forEach(config => {
       expect(device.rainRate.service.characteristics[0].value).to.equal("Rain Rate 0.0 in/hour");
       expect(device.eventRain.service.characteristics[0].value).to.equal("Rain Event Total 0.4 in");
       expect(device.hourlyRain.service.characteristics[0].value).to.equal("Rain Hourly Total 0.5 in");
+      expect(device.last24hRain.service.characteristics[0].value).to.equal("Rain Last 24h Total 0.5 in");
       expect(device.dailyRain.service.characteristics[0].value).to.equal("Rain Daily Total 0.0 in");
       expect(device.weeklyRain.service.characteristics[0].value).to.equal("Rain Weekly Total 0.3 in");
       expect(device.monthlyRain.service.characteristics[0].value).to.equal("Rain Monthly Total 0.8 in");
@@ -154,6 +157,7 @@ configs.forEach(config => {
 
       // optional
       expect(device.rainRate).to.be.undefined;
+      expect(device.last24hRain).to.be.undefined;
       expect(device.totalRain).to.be.undefined;
       done();
     });
@@ -173,8 +177,13 @@ configs.forEach(config => {
       platform.config.nameOverrides[11] = {"key": "WH65:rainDailyTotal", "value": "Test Rain Daily Total Name"};
       platform.config.nameOverrides[12] = {"key": "WH65:rainWeeklyTotal", "value": "Test Rain Weekly Total Name"};
       platform.config.nameOverrides[13] = {"key": "WH65:rainMonthlyTotal", "value": "Test Rain Monthly Total Name"};
-      platform.config.nameOverrides[14] = {"key": "WH65:rainYearlyTotal", "value": "Test Rain Yearly Total Name"};
-      platform.config.nameOverrides[15] = {"key": "WH65:rainTotal", "value": "Test Rain Total Name"};
+      platform.config.nameOverrides[14] = {"key": "WH65:rainHourlyTotal", "value": "Test Rain Hourly Total Name"};
+      platform.config.nameOverrides[15] = {"key": "WH65:rainLast24hTotal", "value": "Test Rain Last 24h Total Name"};
+      platform.config.nameOverrides[16] = {"key": "WH65:rainDailyTotal", "value": "Test Rain Daily Total Name"};
+      platform.config.nameOverrides[17] = {"key": "WH65:rainWeeklyTotal", "value": "Test Rain Weekly Total Name"};
+      platform.config.nameOverrides[18] = {"key": "WH65:rainMonthlyTotal", "value": "Test Rain Monthly Total Name"};
+      platform.config.nameOverrides[19] = {"key": "WH65:rainYearlyTotal", "value": "Test Rain Yearly Total Name"};
+      platform.config.nameOverrides[20] = {"key": "WH65:rainTotal", "value": "Test Rain Total Name"};
 
       device = new WH65(platform, accessory);
 
@@ -189,6 +198,7 @@ configs.forEach(config => {
       expect(device.rainRate.service.characteristics[0].value).to.equal("Test Rain Rate Name");
       expect(device.eventRain.service.characteristics[0].value).to.equal("Test Rain Event Total Name");
       expect(device.hourlyRain.service.characteristics[0].value).to.equal("Test Rain Hourly Total Name");
+      expect(device.last24hRain.service.characteristics[0].value).to.equal("Test Rain Last 24h Total Name");
       expect(device.dailyRain.service.characteristics[0].value).to.equal("Test Rain Daily Total Name");
       expect(device.weeklyRain.service.characteristics[0].value).to.equal("Test Rain Weekly Total Name");
       expect(device.monthlyRain.service.characteristics[0].value).to.equal("Test Rain Monthly Total Name");
@@ -209,6 +219,7 @@ configs.forEach(config => {
       platform.config.hidden["rainRate"] = true;
       platform.config.hidden["rainEventTotal"] = true;
       platform.config.hidden["rainHourlyTotal"] = true;
+      platform.config.hidden["rainLast24hTotal"] = true;
       platform.config.hidden["rainDailyTotal"] = true;
       platform.config.hidden["rainWeeklyTotal"] = true;
       platform.config.hidden["rainMonthlyTotal"] = true;
@@ -228,6 +239,7 @@ configs.forEach(config => {
       expect(device.rainRate).to.be.undefined;
       expect(device.eventRain).to.be.undefined;
       expect(device.hourlyRain).to.be.undefined;
+      expect(device.last24hRain).to.be.undefined;
       expect(device.dailyRain).to.be.undefined;
       expect(device.weeklyRain).to.be.undefined;
       expect(device.monthlyRain).to.be.undefined;
@@ -248,6 +260,7 @@ configs.forEach(config => {
       platform.config.hidden["WH65:rainRate"] = true;
       platform.config.hidden["WH65:rainEventTotal"] = true;
       platform.config.hidden["WH65:rainHourlyTotal"] = true;
+      platform.config.hidden["WH65:rainLast24hTotal"] = true;
       platform.config.hidden["WH65:rainDailyTotal"] = true;
       platform.config.hidden["WH65:rainWeeklyTotal"] = true;
       platform.config.hidden["WH65:rainMonthlyTotal"] = true;
@@ -267,6 +280,7 @@ configs.forEach(config => {
       expect(device.rainRate).to.be.undefined;
       expect(device.eventRain).to.be.undefined;
       expect(device.hourlyRain).to.be.undefined;
+      expect(device.last24hRain).to.be.undefined;
       expect(device.dailyRain).to.be.undefined;
       expect(device.weeklyRain).to.be.undefined;
       expect(device.monthlyRain).to.be.undefined;